Control.Monad.Class.MonadTime.SI

Documentation

class Monad m => MonadTime (m :: Type -> Type) where

Methods

getCurrentTime :: m UTCTime

class MonadMonotonicTimeNSec m => MonadMonotonicTime m where

Minimal complete definition

Nothing

Methods

getMonotonicTime :: m Time

default getMonotonicTime :: m Time

DiffTime and its action on Time

newtype Time

A point in time in a monotonic clock.

The epoch for this clock is arbitrary and does not correspond to any wall clock or calendar, and is not guaranteed to be the same epoch across program runs. It is represented as the DiffTime from this arbitrary epoch.

Constructors

Time DiffTime

diffTime :: Time -> Time -> DiffTime

The time duration between two points in time (positive or negative).

addTime :: DiffTime -> Time -> Time infixr 9

Add a duration to a point in time, giving another time.

data DiffTime

This is a length of time, as measured by a clock. Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: DiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

NominalTime and its action on UTCTime

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

diffUTCTime :: UTCTime -> UTCTime -> NominalDiffTime

diffUTCTime a b = a - b

addUTCTime :: NominalDiffTime -> UTCTime -> UTCTime

addUTCTime a b = a + b

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.