Control.Monad.Class.MonadTimer.SI

Type classes

class (MonadDelay m, MonadMonotonicTime m) => MonadDelay m where

Methods

threadDelay :: DiffTime -> m ()

class (MonadTimer m, MonadMonotonicTime m) => MonadTimer m where

Methods

registerDelay :: DiffTime -> m (TVar m Bool)

A register delay function which safe on 32-bit systems.

registerDelayCancellable :: DiffTime -> m (STM m TimeoutState, m ())

A cancellable register delay which is safe on 32-bit systems and efficient for delays smaller than what Int can represent (especially on systems which support native timer manager).

timeout :: DiffTime -> m a -> m (Maybe a)

A timeout function.

TODO: IO instance is not safe on 32-bit systems.

Auxiliary functions

diffTimeToMicrosecondsAsInt :: DiffTime -> Int

Convert DiffTime in seconds to microseconds represented by an Int.

Note that on 32bit systems it can only represent `2^31-1` seconds, which is only ~35 minutes.

It doesn't prevent under- or overflows; when assertions are on it will thrown an assertion exception.

microsecondsAsIntToDiffTime :: Int -> DiffTime

Convert time in microseconds in DiffTime (measured in seconds).

Re-exports

data DiffTime

This is a length of time, as measured by a clock. Conversion functions such as fromInteger and realToFrac will treat it as seconds. For example, (0.010 :: DiffTime) corresponds to 10 milliseconds.

It has a precision of one picosecond (= 10^-12 s). Enumeration functions will treat it as picoseconds.

class MonadThread m => MonadFork (m :: Type -> Type)

Minimal complete definition

forkIO, forkOn, forkIOWithUnmask, forkFinally, throwTo, yield

class MonadMonotonicTimeNSec m => MonadMonotonicTime m

class Monad m => MonadTime (m :: Type -> Type)

Minimal complete definition

getCurrentTime

data TimeoutState

State of a timeout: pending, fired or cancelled.

Constructors

TimeoutPending

TimeoutFired

TimeoutCancelled

Default implementations

defaultRegisterDelay :: forall m timeout. (MonadFork m, MonadMonotonicTime m, MonadSTM m) => NewTimeout m timeout -> AwaitTimeout m timeout -> DiffTime -> m (TVar m Bool)

A default implementation of registerDelay which supports delays longer then Int; this is especially important on 32-bit systems where maximum delay expressed in microseconds is around 35 minutes.

defaultRegisterDelayCancellable :: forall m timeout. (MonadFork m, MonadMonotonicTime m, MonadSTM m) => NewTimeout m timeout -> ReadTimeout m timeout -> CancelTimeout m timeout -> AwaitTimeout m timeout -> DiffTime -> m (STM m TimeoutState, m ())

A cancellable register delay which is safe on 32-bit systems and efficient for delays smaller than what Int can represent (especially on systems which support native timer manager).